Add estimator module to fix circular import error
