+++ nimare/meta/ibma.py
@@ -12,7 +12,7 @@ from nilearn.image import concat_imgs, resample_to_img
 from nilearn.input_data import NiftiMasker
 from nilearn.mass_univariate import permuted_ols
 
-from nimare.base import Estimator
+from nimare.estimator import Estimator
 from nimare.transforms import p_to_z, t_to_z
 from nimare.utils import _boolean_unmask, _check_ncores, get_masker
